+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Fixed
 - ``evc.current_path`` is deleted when an error with TAG type is raised.
 - Link up from UNI will deploy correctly an EVC when it does not have a path.
 - Fixed Path ``choose_vlans`` to be all or nothing, if a path link fails to allocate a vlan, it'll release the allocated vlans.
+- Validate paths from ``pathfinder`` with interface couples, in and out interfaces.
 
 Added
 =====